Change title 'sandwich' text
##Changed
You're viewing an update of

##To
You're viewing the edit history of

##Reason
'An update of' doesn't really make sense - this introduces a new term, 'edit history', but I'd like to see how that tests.

--

##Changed
published on

##To
this update was made on

##Reason
Gives context to the date. 'Published on' is ambiguous as could refer to the content as a whole, ie look like it's the 'first published' date.